+++ src/main/java/org/sopt/makers/operation/dto/attendance/AttendRequestDTO.java
@@ -2,6 +2,6 @@ package org.sopt.makers.operation.dto.attendance;
 
 public record AttendRequestDTO(
         Long subLectureId,
-        Long code
+        String code
 ) {
 }
