+++ src/main/java/org/sopt/makers/operation/service/MemberService.java
@@ -2,6 +2,7 @@ package org.sopt.makers.operation.service;
 
 import org.sopt.makers.operation.dto.attendance.AttendanceTotalResponseDTO;
 import org.sopt.makers.operation.dto.member.MemberListGetResponse;
+import org.sopt.makers.operation.dto.member.MemberScoreGetResponse;
 import org.sopt.makers.operation.entity.Member;
 import org.sopt.makers.operation.entity.Part;
 
@@ -9,7 +10,7 @@ import java.util.Optional;
 import java.util.List;
 
 public interface MemberService {
-    Optional<Member> confirmMember(Long playGroundId);
     List<MemberListGetResponse> getMemberList(Part part, int generation);
-    AttendanceTotalResponseDTO getMemberTotalAttendance(Member member);
+    AttendanceTotalResponseDTO getMemberTotalAttendance(Long playGroundId);
+    MemberScoreGetResponse getMemberScore(Long playGroundId);
 }
